+++ packages/core/src/entity/EntityStoreBase.ts
@@ -1,12 +1,15 @@
 import type { Asset, ChainModifiers, Entry, UnresolvedLink } from 'contentful';
 
 import { get } from '../utils/get';
+import { isLink } from '../utils/isLink';
+import { isDeepPath, parseDataSourcePathIntoFieldset } from '@/utils/pathSchema';
+// import { transformAssetFileToUrl } from './value-transformers';
 
 /**
  * Base Store for entities
  * Can be extened for the different loading behaviours (editor, production, ..)
  */
-export class EntityStoreBase {
+export abstract class EntityStoreBase {
   protected locale: string;
   protected entryMap = new Map<string, Entry>();
   protected assetMap = new Map<string, Asset>();
@@ -27,6 +30,125 @@ export class EntityStoreBase {
     this.addEntity(entity);
   }
 
+  public getEntryOrAsset(
+    linkOrEntryOrAsset: UnresolvedLink<'Entry' | 'Asset'> | Asset | Entry,
+    // link: UnresolvedLink<'Entry' | 'Asset'>,
+    path: string,
+  ): Entry | Asset | undefined {
+    const resolveFieldset = (
+      unresolvedFieldset: Array<[null, string, string?]>,
+      headEntry: Entry | Asset,
+    ) => {
+      const resolvedFieldset: Array<[Entry | Asset, string, string?]> = [];
+      let entityToResolveFieldsFrom: Entry | Asset = headEntry;
+      for (let i = 0; i < unresolvedFieldset.length; i++) {
+        const isLeaf = i === unresolvedFieldset.length - 1; // with last row, we are not expecting a link, but a value
+        const row = unresolvedFieldset[i];
+        const [, field, _localeQualifier] = row;
+        if (!entityToResolveFieldsFrom) {
+          throw new Error(
+            `Logic Error: Cannot resolve field ${field} of a fieldset as there is no entity to resolve it from.`,
+          );
+        }
+        if (isLeaf) {
+          resolvedFieldset.push([entityToResolveFieldsFrom, field, _localeQualifier]);
+          break;
+        }
+
+        const fieldValue = get<string | UnresolvedLink<'Entry' | 'Asset'>>(
+          entityToResolveFieldsFrom,
+          ['fields', field],
+        );
+
+        if (undefined === fieldValue) {
+          return {
+            resolvedFieldset,
+            isFullyResolved: false,
+            reason: `Cannot resolve field Link<${entityToResolveFieldsFrom.sys.type}>(sys.id=${entityToResolveFieldsFrom.sys.id}).fields[${field}] as field value is not defined`,
+          };
+        } else if (isLink(fieldValue)) {
+          const entity = this.getEntityFromLink(fieldValue);
+          if (entity === undefined) {
+            throw new Error(
+              `Logic Error: Broken Precondition [by the time resolution of deep path happens all referents should be in EntityStore]: Cannot resolve field ${field} of a fieldset row [${JSON.stringify(
+                row,
+              )}] as linked entity not found in the EntityStore. ${JSON.stringify({
+                link: fieldValue,
+              })}`,
+            );
+          }
+          resolvedFieldset.push([entityToResolveFieldsFrom, field, _localeQualifier]);
+          entityToResolveFieldsFrom = entity; // we move up
+        } else {
+          // TODO: Eg. when someone changed the schema and the field is not a link anymore, what should we return then?
+          throw new Error(
+            `LogicError: Invalid value of a field we consider a reference field. Cannot resolve field ${field} of a fieldset as it is not a link, neither undefined.`,
+          );
+        }
+      }
+      return {
+        resolvedFieldset,
+        isFullyResolved: true,
+      };
+    };
+
+    if (isDeepPath(path)) {
+      const headEntity = isLink(linkOrEntryOrAsset)
+        ? this.getEntityFromLink(linkOrEntryOrAsset)
+        : (linkOrEntryOrAsset as Entry | Asset);
+
+      if (undefined === headEntity) {
+        return;
+      }
+      const unresolvedFieldset = parseDataSourcePathIntoFieldset(path);
+
+      // The purpose here is to take this intermediate representation of the deep-path
+      // and to follow the links to the leaf-entity and field
+      // in case we can't follow till the end, we should signal that there was null-reference in the path
+      const { resolvedFieldset, isFullyResolved, reason } = resolveFieldset(
+        unresolvedFieldset,
+        headEntity,
+      );
+      if (!isFullyResolved) {
+        reason &&
+          console.debug(
+            `[exp-builder.sdk::EntityStoreBased::getValueDeep()] Deep path wasn't resolved till leaf node, falling back to undefined, because: ${reason}`,
+          );
+        return undefined;
+      }
+      // const [leafEntity, field /* localeQualifier */] = resolvedFieldset[resolvedFieldset.length - 1];
+      const [leafEntity] = resolvedFieldset[resolvedFieldset.length - 1];
+      return leafEntity;
+      // const fieldValue = get<string>(leafEntity, ['fields', field]); // is allowed to be undefined (when non-required field not set; or even when field does NOT exist on the type)
+      // return transformAssetFileToUrl(fieldValue);
+    } else {
+      let entity: Entry | Asset;
+      if (isLink(linkOrEntryOrAsset)) {
+        const resolvedEntity =
+          linkOrEntryOrAsset.sys.linkType === 'Entry'
+            ? this.entryMap.get(linkOrEntryOrAsset.sys.id)
+            : this.assetMap.get(linkOrEntryOrAsset.sys.id);
+        if (!resolvedEntity || resolvedEntity.sys.type !== linkOrEntryOrAsset.sys.linkType) {
+          console.warn(
+            `Experience references unresolved entity: ${JSON.stringify(linkOrEntryOrAsset)}`,
+          );
+          return;
+        }
+        entity = resolvedEntity;
+      } else {
+        // We already have the complete entity in preview & delivery (resolved by the CMA client)
+        entity = linkOrEntryOrAsset;
+      }
+      return entity;
+    }
+  }
+
+  /**
+   * @deprecated in the base class this should be simply an abstract method
+   * @param entityLink
+   * @param path
+   * @returns
+   */
   public getValue(
     entityLink: UnresolvedLink<'Entry' | 'Asset'>,
     path: string[],
@@ -44,30 +166,43 @@ export class EntityStoreBase {
     return get<string>(entity, path);
   }
 
-  public getEntryOrAsset(entityLinkOrEntity: UnresolvedLink<'Entry' | 'Asset'> | Entry | Asset) {
-    const isLink = (
-      entity: typeof entityLinkOrEntity,
-    ): entity is UnresolvedLink<'Entry' | 'Asset'> => entityLinkOrEntity.sys.type === 'Link';
+  // public getEntryOrAsset(entityLinkOrEntity: UnresolvedLink<'Entry' | 'Asset'> | Entry | Asset) {
+  //   const isLink = (
+  //     entity: typeof entityLinkOrEntity,
+  //   ): entity is UnresolvedLink<'Entry' | 'Asset'> => entityLinkOrEntity.sys.type === 'Link';
 
-    let entity: Entry | Asset;
-    if (isLink(entityLinkOrEntity)) {
-      const resolvedEntity =
-        entityLinkOrEntity.sys.linkType === 'Entry'
-          ? this.entryMap.get(entityLinkOrEntity.sys.id)
-          : this.assetMap.get(entityLinkOrEntity.sys.id);
+  //   let entity: Entry | Asset;
+  //   if (isLink(entityLinkOrEntity)) {
+  //     const resolvedEntity =
+  //       entityLinkOrEntity.sys.linkType === 'Entry'
+  //         ? this.entryMap.get(entityLinkOrEntity.sys.id)
+  //         : this.assetMap.get(entityLinkOrEntity.sys.id);
 
-      if (!resolvedEntity || resolvedEntity.sys.type !== entityLinkOrEntity.sys.linkType) {
-        console.warn(
-          `Experience references unresolved entity: ${JSON.stringify(entityLinkOrEntity)}`,
-        );
-        return;
-      }
-      entity = resolvedEntity;
-    } else {
-      // We already have the complete entity in preview & delivery (resolved by the CMA client)
-      entity = entityLinkOrEntity;
+  //     if (!resolvedEntity || resolvedEntity.sys.type !== entityLinkOrEntity.sys.linkType) {
+  //       console.warn(
+  //         `Experience references unresolved entity: ${JSON.stringify(entityLinkOrEntity)}`,
+  //       );
+  //       return;
+  //     }
+  //     entity = resolvedEntity;
+  //   } else {
+  //     // We already have the complete entity in preview & delivery (resolved by the CMA client)
+  //     entity = entityLinkOrEntity;
+  //   }
+  //   return entity;
+  // }
+
+  public getEntityFromLink(link: UnresolvedLink<'Entry' | 'Asset'>): Asset | Entry | undefined {
+    const resolvedEntity =
+      link.sys.linkType === 'Entry'
+        ? this.entryMap.get(link.sys.id)
+        : this.assetMap.get(link.sys.id);
+
+    if (!resolvedEntity || resolvedEntity.sys.type !== link.sys.linkType) {
+      console.warn(`Experience references unresolved entity: ${JSON.stringify(link)}`);
+      return;
     }
-    return entity;
+    return resolvedEntity;
   }
 
   protected getEntitiesFromMap(type: 'Entry' | 'Asset', ids: string[]) {