+++ packages/experience-builder-sdk/src/hooks/useFetchExperience.test.ts
@@ -15,14 +15,17 @@ let clientMock: ContentfulClientApi<undefined>;
 describe('useFetchExperience', () => {
   beforeEach(() => {
     clientMock = {
-      getEntries: jest.fn().mockImplementation((data) => {
-        if ('sys.id[in]' in data) {
-          return Promise.resolve({ items: entries });
-        }
-
+      getEntries: jest.fn().mockImplementation((_query) => {
+        // { content_type: 'layout', locale: 'en-US', 'fields.slug': 'hello-world' }
         return Promise.resolve({ items: [compositionEntry] });
       }),
       getAssets: jest.fn().mockResolvedValue({ items: assets }),
+      withoutLinkResolution: {
+        getEntries: jest.fn().mockImplementation((_query) => {
+          // { "sys.id[in]": [ "entry1", "entry2", ], locale: "en-US", }
+          return Promise.resolve({ items: entries });
+        }),
+      },
     } as unknown as ContentfulClientApi<undefined>;
   });
 
@@ -51,6 +54,7 @@ describe('useFetchExperience', () => {
     expect(store).toEqual({
       experience: undefined,
       isFetching: false,
+      error: undefined,
       fetchBySlug: store.fetchBySlug,
       fetchById: store.fetchById,
     });
@@ -68,7 +72,8 @@ describe('useFetchExperience', () => {
         locale: localeCode,
       });
 
-      expect(experience?.entityStore).toMatchObject(entityStore);
+      expect(experience).toHaveProperty('entityStore');
+      expect(experience!.entityStore).toMatchObject(entityStore);
     });
 
     expect(clientMock.getEntries).toHaveBeenNthCalledWith(1, {
@@ -77,7 +82,7 @@ describe('useFetchExperience', () => {
       locale: localeCode,
     });
 
-    expect(clientMock.getEntries).toHaveBeenNthCalledWith(2, {
+    expect(clientMock.withoutLinkResolution.getEntries).toHaveBeenNthCalledWith(1, {
       'sys.id[in]': entries.map((entry) => entry.sys.id),
       locale: localeCode,
     });
@@ -105,6 +110,7 @@ describe('useFetchExperience', () => {
     expect(store).toEqual({
       experience: undefined,
       isFetching: false,
+      error: undefined,
       fetchBySlug: store.fetchBySlug,
       fetchById: store.fetchById,
     });
@@ -122,7 +128,8 @@ describe('useFetchExperience', () => {
         locale: localeCode,
       });
 
-      expect(experience?.entityStore).toMatchObject(entityStore);
+      expect(experience).toHaveProperty('entityStore');
+      expect(experience!.entityStore).toMatchObject(entityStore);
     });
 
     expect(clientMock.getEntries).toHaveBeenNthCalledWith(1, {
@@ -131,7 +138,7 @@ describe('useFetchExperience', () => {
       locale: localeCode,
     });
 
-    expect(clientMock.getEntries).toHaveBeenNthCalledWith(2, {
+    expect(clientMock.withoutLinkResolution.getEntries).toHaveBeenNthCalledWith(1, {
       'sys.id[in]': entries.map((entry) => entry.sys.id),
       locale: localeCode,
     });